Improve slabs
This commit does a lot.

First, it makes sure that dc works with the slabs, even though it cleans
with them now (I forgot that when I introduced slabs).

Second, I removed a place where I used the vm slabs for the bc parser
and made it use a special slab in the parser itself.

Then, I made slab vector code to do the handling, in order to remove
problems. I still expect that AFL++ will find crashes, though...

Signed-off-by: Gavin Howard <gavin@yzena.com>