Make bc_num_irand() vastly more efficient
This removes all math in it, and just has it generate hardware integers
up to BC_BASE until at the last limb of the limit. If the limit's limb
is one, we skip; otherwise, we generate with the limb as the limit.

From taking a minute to run the irand_long benchmark to 17 seconds. The
change on the irand_short benchmark was minor enough to not count, but
since there was no change there, the long benchmark matters.

Signed-off-by: Gavin Howard <gavin@yzena.com>