Add the ability for users to redefine keywords
This was an idea from Stefan Esser of FreeBSD. At first, I did not like
it because it seemed like it would require a deep change to the parser,
to check everywhere a name could be used.

Fortunately, I came up with the idea of using an array of booleans in
BcVm to allow the parser to mark when a keyword has been redefined,
which can only happen in one place: defining functions. This made the
change to the parser simple at the expense of another simple change to
the bc lexer (literally only one line).

Because of the use of this feature in allowing users to make my bc parse
GNU bc-compatible scripts, as well as interoperate with scripts from
other bc's, I am putting this change in.

Signed-off-by: Gavin Howard <gavin@yzena.com>