Add information in the bc manual for the keyword redefinition
The keyword definition feature was committed in the previous two
commits. This commit makes the bc manual match.

Oh, part of the information is a blurb at the top saying that if this bc
fails to parse a script for another bc, to define the environment
variable BC_REDEFINE_KEYWORDS to 1 and try again. This is to make it
easy for frustrated users to find a solution to their problem.

Yes, it is hidden behind an environment variable. This is because I
*don't* want this behavior to be default.

Signed-off-by: Gavin Howard <gavin@yzena.com>