Change the keyword redefinition infrastructure
This changes it to use a command-line option. This is best because if a
user encounters a problem; they *know* they are using my bc and can add
the command-line option.

This is also better because it allows the keyword to be used as variable
and array names as well as function names, without editing the script.

This has the double advantage that I can undo the change to the parser,
though I need to keep the one line change to the lexer.

The option is `-r` and `--redefine`, and they take an argument, which is
the keyword to redefine.

All docs and tests have been updated.

Signed-off-by: Gavin Howard <gavin@yzena.com>