Fix a crash found by AFL++
This crash was caused by a value needing to be used after an assignment
to a string, which never needed a value used again before because only
dc did it. So this commit changes it so another string is pushed, if
necessary.

This commit also adds a test for it. It's not an error test; it's just a
normal test added to bc's strings test.

Signed-off-by: Gavin Howard <gavin@yzena.com>