Fix a crash in dc found by AFL++
This crash was caused by cleaning constants in other "functions"
(strings). It happened when a string of exactly the same text was used
again, which means it already had a function ready for it.
Unfortunately, since that function had stored constants, it expected
them to be there. They were not, as they had been cleaned.

But they should not have been cleaned; only the constants in main should
have been.

I found the problem: dc was still using the main const slab for all
constants, but it should not have been. This commit changes that.

Signed-off-by: Gavin Howard <gavin@yzena.com>