Parallelize tests/all.sh
This is for FreeBSD, which runs the tests as part of their test suite.
With the history tests, the test suite takes 100-120 seconds to run.
With this change, it gets it under 30 seconds.

The parallelism is...crude. What it does is spawn as many tests into the
background as posssible and then waits on them all, exiting with an
error if any of them failed. There is no special sauce with the GNU
parallel command because that's not portable, and I probably can't get a
guarantee that it will even exist on FreeBSD systems running the test.

Signed-off-by: Gavin Howard <gavin@yzena.com>