Force Windows to use binary mode for files
It turns out that if I do not, it silently replaces CRLF with LF and
tells bc that it read less bytes than it did. [1] Wat?!

So, I have to force it to use binary mode instead of text mode so that
it will report the true numbers to me. Sigh...I hate Windows.

[1]: https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/read

Signed-off-by: Gavin Howard <gavin@yzena.com>