Add a command-line option for printing leading zeros
This was also requested by depler on GitHub. It took a bit of
refactoring, but it actually wasn't bad.

Later, I will add functions to the extended math library to either print
the leading zero for certain numbers, or not. To do that, I am going to
add builtin functions to query the global state. It makes sense to do so
since I already have other builtin functions for global state.

Signed-off-by: Gavin Howard <gavin@yzena.com>