Fix a bug found by the test in the previous commit
It turns that I was not properly ending if statements for the beginning
of functions. If I had an if statement that did not have an else right
before defining a function, bc thought that it was invalid. So I added a
specific check.

Thank goodness I now have a test for it.

Signed-off-by: Gavin Howard <gavin@yzena.com>