Fix interaction between signals and globals
I was thinking about setjmp() and longjmp(), and I realized that they
are still dangerous in the presence of global variables. bc has a big
one: BcVm vm.

This commit basically locks signals wherever something directly inside
of BcVm is changed. (Things pointed to do not need to be locked or
already are.) This makes mutating BcVm safe even in the presence of
signals.

Signed-off-by: Gavin Howard <gavin@yzena.com>