Fix the bug mentioned in the last commit
This bug was a little bit of an easy one to find and a doozy to fix. The
reason for that is because reading expressions from the command-line is
kind of its own special case, so I had to do a little bit of a redesign.

The redesign wasn't that big, however; I reused the same technique I
used for fixing a similar problem with stdin. I also reused the same
global buffer in BcVm because stdin won't ever use it at the same time.
I even used kind of the same code.

But the idea is the same: have a function the lexer can call when
there is an unfinished comment or string. In fact, I need to add a test
for both of those cases, as well as a test for a multiline string.

Signed-off-by: Gavin Howard <gavin@yzena.com>