Fix a problem found with the fix in last commit
The problem was that if a backslash was at the end of the line, it
correctly asked for more, but if it didn't get any, it happily just
quit. No bueno.

So if it requests more and doesn't get it, we should attempt to process
it. This will flush out any errors.

Signed-off-by: Gavin Howard <gavin@yzena.com>