Implement support for editline and readline
This was requested by an important FreeBSD user. This was an exploration
of the idea to just use editline and readline. I wasn't sure it was
possible. Turns out, it is.

What is not working: history in editline does not seem to work.
Everything else seems to work. Tests still pass as well.

Signed-off-by: Gavin Howard <gavin@yzena.com>