Fix style
This style fixing is to separate out local variables that are pointers
into separate lines, instead of being mashed together. This is an
artifact of the terrible style that was imposed on bc by toybox, but
since I don't maintain that code, I don't care anymore.

Also, I'm trying again to implement automatic code formatting.

Signed-off-by: Gavin Howard <gavin@yzena.com>