Add a warning to configure.sh about braindead POSIX locales
This warning is there to warn users that locales will be installed
outside of $PREFIX, so it first detects if the locales are in $PREFIX or
if the default $PREFIX is used. If so, it skips the warning. Otherwise,
it displays it.

This is a change that was sort of requested by a user in GitHub issue
number 53.

Signed-off-by: Gavin Howard <gavin@yzena.com>