Make bcl work with threads and not signals
Unfortunately, I couldn't have both threads *and* signals, and I feel
like threads would be more widely applicable. Plus, I want them for a
project I want to work on.

The reason they don't mix is because I had to use thread-specific data,
and there is no safe way to get that data in a signal handler. I also
thought that it was too much complexity, so I just removed it.

Signed-off-by: Gavin Howard <gavin@yzena.com>