Hook the bcl test into the Valgrind testing infrastructure
This should have been done long ago; it has already found memory leaks
and memory bugs. Whoops.

I found this because I realized that I forgot to free the BcVm on each
thread, and I realized that the release script should have caught that
if it was properly hooked in.

Signed-off-by: Gavin Howard <gavin@yzena.com>