Attempt to fix a build bug
This appeared to be caused by c36b91b024e, which says that it is trying
to fix an issue on the FreeBSD port. However, half of that commit has
already been removed, and when I remove the other half, the port builds
fine still.

It looks like that commit may have been entirely unnecessary due to
later changes, but it broke cross-compiling.

This commit removes the rest of that commit. This may not work, but I'm
going to try it.

Signed-off-by: Gavin Howard <gavin@yzena.com>