Fix a crash found by AFL
This crash was caused by a "suggestion" (more like an order) given to me
by the BusyBox maintainer way back when. He had wanted me to make it so
strings, constants, and such things were freed once code had been
executed.

This clean-up code has given me a lot of trouble since.

Well, the crash was caused by redefining a function after it had been
run once. There were a few other requirements, such as having the
function access a non-auto variable and put a string into it by the
asciify() function. Redefining the function meant that the strings
stored in the function went away, and since the variable was assigned a
string that only existed in the function, when it tried to access the
variable, the assert in bc_vec_item() fired.

(Yes, this means that there was a memory safety bug.)

The solution was to make strings and constants be global to the program,
like functions, variables, and arrays. Also, I had to make them use
the vector maps, so that multiple instances of the same string would not
be used. (Well, multiple for each type.)

In turn, this simplified the handling of both. Strings didn't need to
set their function anymore, for example. It also reduces memory usage on
duplicated data.

It also simplifies the use of slab vectors; there's only one now, and it
stores strings of all types. This, in turn, required reducing what could
be cleaned up once code had executed. I don't mind, though; as I said
above, this code has been some of the most problematic in the entire
repo.

Signed-off-by: Gavin Howard <gavin@yzena.com>