Fix a performance problem
The problem is that newlines were not being eaten at the end of
statements, so there was extra time spent starting execution of
non-existent code. This wasn't a problem in the previous versions
because the parse function was called as many times as necessary, but in
this one, it is called only once per execution try.

Signed-off-by: Gavin Howard <gavin@yzena.com>