Fix a crash found by AFL
This crash was caused by mishandling of `bc` parse flags: the array flag
should have been cleared for all single-argument built-in functions
besides length, and it was not. Thus, an array argument was allowed
through, causing an out-of-bounds access at least on `asciify()`, and
maybe in more cases.

The solution was to explicitly clear the array flag.

Signed-off-by: Gavin Howard <gavin@yzena.com>