Add the is_number() and is_string() built-in functions
These are `u` and `t` (the second letters of the second word),
respectively, in `dc`.

This commit includes all of the tests, including moving a now-valid `dc`
error test into the regular test suite.

This is also my first public commit with my new Curve 25519 key. I
wanted to switch to ECC for various reasons.

Signed-off-by: Gavin Howard <gavin@yzena.com>