Fix a crash found by AFL
This crash was caused by sending string arguments to
bc_program_asciifyNum(), which shouldn't happen. After all, it's got
"Num" in the name. So I added an if statement for it.

This was a crash from a bug in new code for this release, so we caught a
bug before release! Yay!

Signed-off-by: Gavin Howard <gavin@yzena.com>