Fix the fatal I/O error with editline on FreeBSD-CURRENT and dwm
This was caused by libedit returning an error when it was just
interrupted. I assumed libedit would be better than that. <facepalm>

I should not have assumed.

Unfortunately, libedit is still a problem because it will forget
everything typed at the prompt. It will also reprint the prompt if I
don't tell it not to. Hence, the extra global boolean in this commit.

Signed-off-by: Gavin Howard <gavin@yzena.com>