Add a test for a bug I found while writing a script
This bug is complicated, but it is caused when an element from an array
is passed as a parameter, and then a later parameter is named the same
as the array whose element was passed in.

I still don't know how to solve it, but this test fails properly while
the bug still exists.

Signed-off-by: Gavin Howard <gavin@yzena.com>