Fix the bug mentioned in the previous commit
So the fix was much easier than I thought. I just had to add an extra
field to BcLoc for the stack index and set it, and then when actually
grabbing the element, I needed to grab the vector at the index.

Adding an element to BcLoc was free, by the way. The reason is that
BcLoc is only used in a union with a BcNum and a BcVec, both of which
are bigger, so adding the extra field doesn't take any more space.

Signed-off-by: Gavin Howard <gavin@yzena.com>