Expand the bug fix from three commits ago to simplify code
The bug fix (commit 2484c9b600) put a stack index into a BcLoc. I
decided to also use this for variables. This allowed me to simplify
function calls in two ways:

* I don't have to check that a parameter that is used has already been
  pushed to.
* I don't have to do anything special to push a parameter.

The last one is especially important. `bc_program_copyToVar()`, the
function that pushes parameters, is complex for bc, and it was *more*
complex before this commit.

Signed-off-by: Gavin Howard <gavin@yzena.com>