Add error checking for args in scripts
This was (sort of) requested by Android. They couldn't update beyond
5.0.1 because I'd changed a test script, and the output was so messy
that it was impossible to see why. It turns out that it was an invalid
argument.

I was told that it would have been better if I didn't have optional
arguments, which is correct. Unfortunately, I have to work with what
I've got, so I did the next best thing.

Signed-off-by: Gavin Howard <gavin@yzena.com>