Fix a build issue with gen/strgen.sh
This boolean can be passed empty, so we need to check that too. However,
I'd still like the test to exist since there is a use that requires a
number.

I also fixed the Makefile to always pass all arguments, and I fixed
strgen.c to respect a `0` argument for remove_tabs. (If the argument
existed at all, tabs were removed. This was different from strgen.sh,
and it was not what I intended.)

Signed-off-by: Gavin D. Howard <gavin@gavinhoward.com>