Fix a bug in root() and cbrt()
This was reported on GitHub. It was caused by a pathological case where
the two numbers to be compared never became equal.

I changed it to add some precision based on the precision of the
argument and the current scale, then to compare the values truncated to
the scale plus change.

Signed-off-by: Gavin D. Howard <gavin@gavinhoward.com>