Update the GDH preferred build again
The Gentoo bug is now fixed, but Clang 16 added a new warning for unsafe
buffer access. This warning is meant for C++ where you *can* have safe
buffer accesses, ones with bounds checks, but in C, it thinks every
buffer access is unsafe, even when they are bounds checked.

This includes argv, when you can check bounds with argc! It's dumb, so I
just turn that warning off.

Signed-off-by: Gavin D. Howard <gavin@gavinhoward.com>