Fix a bug found by a user
This bug is that an error can happen when flushing the streams on exit,
and in that case, `bc`/`dc` will jump out, but there won't be any
jmp_buf stack, so crash.

I can't reproduce it, but I'm pretty sure this is the fix, and I will
confirm with the reporter.

Signed-off-by: Gavin D. Howard <gavin@gavinhoward.com>