Improve the fix from last commit
It turns out, obviously, that ignoring errors on flush will prevent any
error message from being printed. This goes against the `bc` standard,
as the same user pointed out to me. Whoops.

The proper fix might be to move the `bc_vm_atexit()` call to *before*
the last `setjmp()` clears. This commit is that fix; I will confirm with
the user.

This reverts last commit as well.

Signed-off-by: Gavin D. Howard <gavin@gavinhoward.com>