Improve the last two commits
The user says that a zero exit status is returned. Whoops.

So I am having `main()` return vm->status instead, which should be set
by a jump.

And I also need to make sure that it is set by the return value of the
calculator mains.

I also need to make sure to grab the return value of the calculator
mains on builds with just one calculator.

Signed-off-by: Gavin D. Howard <gavin@gavinhoward.com>