Fix a bug on the macOS build
This bug is actually a workaround for libedit, which for some reason,
returns a length that isn't -1 after handling SIGWINCH twice.

But whatever the case, this fix works on Linux and macOS, so I'm going
with it.

The fix is to loop around `el_gets()` if *either* len is -1 or errno is
EINTR, instead of requiring both.

Signed-off-by: Gavin D. Howard <gavin@gavinhoward.com>