Fix a crash found by AFL++
The problem was actually in the parser and manifested itself when
executing because the parser should have rejected invalid code.

The problem is that arrays were allowed to be part of bigger
expressions, but that's not allowed; they should always be alone, and
they should always be function parameters.

This commit also adds a minimized AFL++ crash test, as well as another
to make sure the changes in this commit are correct in another corner
case.

This is not a memory bug, insofar as I can tell. If someone figures out
otherwise, though, I correct myself.

Signed-off-by: Gavin D. Howard <gavin@gavinhoward.com>