Fix the leaks from the last three commits for real
So the fixes I made were not great since errors could make them bypass
all of the freeing stuff.

And where is the freeing stuff? In `bc_vm_atexit()`.

The real problem was that `bc_vm_atexit()` was called in `bc_vm_boot()`,
not in `main()` after all of the jumping ended.

So I undid the changes from the past three commits and moved the call to
`bc_vm_atexit()` to `main()`.

Signed-off-by: Gavin D. Howard <gavin@gavinhoward.com>