Let libedit handle terminal size changes.
Previously, we would catch SIGWINCH and call el_resize() from the signal handler.  This is unsafe and led to strange behavior such as terminating on second resize.  The simplest solution is to let libedit handle SIGWINCH itself.

This reverts 56bb18255a24 and 89d6c3451a60 and removes all traces of SIGWINCH from bc itself, and instead sets the EL_SIGNAL flag on the editline context, which causes libedit to detect and handle terminal size changes internally.