Fix a crash in dc
This was found by a user, who ran dc like this:

```
>>> *
>>> f
```

After the first, line, a runtime error happened. However, *before* that
error, a result was pushed onto the stack. That result was still on the
stack when the stack was printed with `f`, but since it was an empty
result, dc crashed.

This commit changes it so the number of non-retired results is stored,
and when dc has an error, those results are popped from the results
stack.

I also kept this behavior in bc to remove the need for one macro
argument, and also to make sure it is properly tested.

Signed-off-by: Gavin D. Howard <gavin@gavinhoward.com>