Implement the entire test suite into Rig
This is the first step in making the test suite work on Windows. And
after this point, Rig is now the official way to develop `bc` and `dc
because it's the only thing that runs the full test suite.

Yes, I actually stopped the POSIX make build from running the whole test
suite. The reason I did that is because some tests were one-offs, and I
don't want to maintain those one-offs in two places. Instead, I put them
all into Rig.

However, most tests are basically plain text files (input for the
calculators) with shell scripts or Rig to marshall them. I kept those in
the POSIX make build by hooking them into the `tests/all.sh` script and
having the POSIX make build use that instead. So POSIX make will still
run the test suite, and most of it, but it doesn't have *all* of it
anymore.

This allowed me to take out some gnarly generation code in the configure
script as well.

I also did some refactoring in the Rig test suite. I also made the `dc`
Easter script an actual script test.

Signed-off-by: Gavin D. Howard <gavin@gavinhoward.com>