Add the maintenance terms
My life has changed drastically recently, and I can no longer afford to
give my time away on this project as I have before. This means that, if
someone finds a bug, wants it fixed, and wants a release, I am no longer
willing to do that without getting compensation.

The maintenance terms list how that will be done and the rates.

I apologize that it has to be this way, but all good things come to an
end.

Signed-off-by: Gavin D. Howard <gavin@gavinhoward.com>